pyproject.toml
[project]
name = "django_todo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "crispy-bootstrap5>=2024.10",
    "django-crispy-forms>=2.3",
    "django>=5.1.2",
]
